refactor: implement targeted on-demand revalidation and disable auth to local dev
