+++ app/admin/page.tsx
@@ -9,15 +9,11 @@ import type { Comic } from '../../types/comic';
 export default async function AdminPage() {
     const session = await auth();
 
-    if (!session) {
-        redirect('/auth/signin');
-    }
+    // if (!session) {
+    //     redirect('/auth/signin');
+    // }
 
     const comics: Comic[] = await getComics();
 
-    return (
-        <>
-            <AdminContent comics={comics} />
-        </>
-    );
+    return <AdminContent comics={comics} />;
 }
