+++ middleware.ts
@@ -5,5 +5,5 @@ export const { auth } = NextAuth(authConfig);
 export default auth;
 
 export const config = {
-    matcher: ["/admin", "/admin/:path*"]
+    matcher: ["/admins", "/admins/:path*"]
 };
